Trigger smoke-test on commits to cf-k8s-networking
- The dont-smoke-for-k8s-deploy-and-test job was previously triggering
on cf-for-k8s-master. It makes more sense to trigger on the
cf-k8s-networking repository because that we will want to validate
changes we make to our code and configuration.
- We are only triggering on changes to the cf-k8s-networking/config
directory because that is the directory that contains our configuration
and image digests that have made it through unit/integration tests.
- Removed cf-k8s-networking-values because it isn't necessary anymore
with the change to the test jobs/image publish jobs, which trigger
explicitly on the controller's directory
468e9566cbc7b29af289f0a52fcca6ebbb10f925

[#172418041](https://www.pivotaltracker.com/story/show/172418041)

Co-authored-by: Tim Downey <tdowney@vmware.com>