ci: Add data plane/control plane uptime test
Welcome to mega commit message. This was a whole branch squashed into
one commit. We decided to be pretty verbose about what happened here,
hopefully this is slightly helpful in the future, since a bunch of
things are introduced.

Here is what happened:
- the data plane test measures the data plane's availability by running
an http get request against an application deployed onto the platform
while an upgrade is running
- the control plane test measures the control plane's availability by
running a map route on the platform and subsequently running http get
requests against the newly mapped route while an upgrade is running
- add a readme that further explains how the measurements are calculated
and more concretely defines our SLOs and SLIs
- add a helpers.sh file in the tasks folder. So far, it only contains
functions to target a cf and a k8s cluster. I figure we will reuse these
things, so might as well pull them out. We can also add other universal
functions to this file instead of redefining them slightly differently
in each task.

The pipeline flow is:
 (triggers on new versions of cf-k8s-networking)
 "create-gke-cluster-and-deploy-cf"
 (cf-install-values triggers)
 "push-apps"
(cf-install-values triggers)
 "upgrade-cf"
 "destroy-cluster"

- install values is the trigger for the chain of jobs, as I
think that needs to be passed through to each job.
- the whole pipeline gets triggered by new versions of cf-k8s-networking
- The destroy cluster does not happen on a trigger.
Create-gke-cluster-and-deploy-cf destroys it if it exists, so I figured,
while in development, we can not destroy it every time while we work out
the kinks. I assume once it is more stable, we can put a trigger on the
destory-cluster job.

[#171202860](https://www.pivotaltracker.com/story/show/171202860)

Co-authored-by: Kauana dos Santos <kdossantos@vmware.com>
Co-authored-by: Nitya Dhanushkodi <ndhanushkodi@pivotal.io>
Co-authored-by: Clay Kauzlaric <ckauzlaric@vmware.com>
Co-authored-by: Josh Russett <jrussett@vmware.com>
Co-authored-by: Christian Ang <angc@vmware.com>