test: Extend waiting time for cf commands in acceptance test
We found that 5 seconds is sometimes not long enough for getting a response from
the cloud controller in our test environments. Since the speed of the CC is not
the point of our tests, we're bumping the timeout from 5 seconds to 30 seconds.

This commit also includes minor cleanup related to the acceptance tests, like
specifying the exact format of the config file for testing explicitly in the README
and correcting the usage string for bin/test_local

[#174425794](https://www.pivotaltracker.com/story/show/174425794)

Signed-off-by: Mikael Manukyan <mmanukyan@vmware.com>
Co-authored-by: Mikael Manukyan <mmanukyan@vmware.com>