scale: all the changes for the scaling pipeline
We've seen the tests run through with P95 of 35s.

- push apps to separate spaces for scaling test
since most of Cloud Controller's resources are scopes by space
and putting 1000 apps in a single space will have performance
consequences for the API

- use proxy app instead of httpbin for scale tests since
proxy is a golang/alpine linux based app which has a lighter footprint
than the python/ubuntu based httpbin image; it should `cf push` a bit
faster

- also add a small sleep between each `cf push` so that we don't spam
the k8s API with map route requests

- use regional cluster for scaling tests since
the k8s control-plane node on the zonal cluster would go offline at times
zonal clusters have a highly available control plane which might help here

- there are two kinds of post-success map-route failures: transient and
persistent. We ignore transient failures; a failure is transient if it
later resolves (later requests return 200). The expectation is that
transient failures happen because we make a request to a gateway while
it is in the midst of being configured by Istio. For persistent
failures, which are failures that never become 200s again, we count the
failures, print the count in the summary, and fail if any occurred.

- use 10 apps per space so that CAPI isn't overwhelmed

[#171678788](https://www.pivotaltracker.com/story/show/171678788)

Co-authored-by: Leah Hanson <lhanson@pivotal.io>
Co-authored-by: Nitya Dhanushkodi <ndhanushkodi@pivotal.io>
Co-authored-by: Alex Standke <astandke@vmware.com>
Co-authored-by: Kauana dos Santos <kdossantos@vmware.com>
Co-authored-by: Mikael Manukyan <mmanukyan@vmware.com>
Co-authored-by: Clay Kauzlaric <ckauzlaric@vmware.com>
Co-authored-by: Tim Downey <tdowney@vmware.com>
Co-authored-by: Mikael Manukyan <mmanukyan@vmware.com>
Co-authored-by: Clay Kauzlaric <ckauzlaric@vmware.com>
Co-authored-by: Leah Hanson <lhanson@pivotal.io>
Co-authored-by: Nitya Dhanushkodi <ndhanushkodi@pivotal.io>
Co-authored-by: Alex Standke <astandke@vmware.com>