scale: give MapRoute more time for cf commands.
Wait will fail the whole test if it triggers, which is probably not the right behavior.
Instead, we want to wait for the timeout and try again. Only the third and final failure/timeout should cause a failure.

This change just doubles the timeout to 30s in hopes that that will make the tests pass.