ci: add temporary contour pipeline
* while we use separate branches to work on contour integration, we
still want to test our work, and provide an environment for PMs to
accept on.
* This is a duplicate of the cf-k8s-pipeline.yml, except that it no
longer interacts with release-candidate. We are opting to work on a
separate branch for now, and merge to develop when we are ready.
* The `contour-controller` tests will not run. I have stubbed out
the jobs and task names, so they should get replaced as we work on
creating a new contour controller.

[#175359407](https://www.pivotaltracker.com/story/show/175359407)